# doc/api.yaml
openapi: 3.0.3
info:
  description: | 
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! You can
    upload your photos directly from your PC, and they will be visible to everyone following you.
  title: WASAPhoto
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
paths:
...
  /users/{UserID}/followed/{OtherUserID}:
    parameters:
      - $ref: "#/components/parameters/UserID"
      - $ref: "#/components/parameters/OtherUserID"
    put:
      tags: ["UserInteraction"]
      operationId: followUser
      summary: Follow another user
      description : "Follow the user if it's not already followed, else do nothing"
      responses:
        '204': { description: "User followed successfully." }
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized"}
        '403': { $ref: "#/components/responses/Forbidden" }
        '404': { $ref: "#/components/responses/ResultNotFound" }
        '409': { description: "The user is already followed, cannot proceed" }
        '500': { $ref: "#/components/responses/InternalServerError" }
    delete:
      tags: ["UserInteraction"]
      operationId: unfollowUser
      summary: Unfollow another user
      description : "Unfollow the user if it's followed, else do nothing"
      responses:
        '204': { description: "User unfollowed successfully." }
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized"}
        '403': { $ref: "#/components/responses/Forbidden" }
        '404': { $ref: "#/components/responses/ResultNotFound" }
        '500': { $ref: "#/components/responses/InternalServerError" }
  /users/{UserID}/banned/{OtherUserID}:
    parameters:
      - $ref: "#/components/parameters/UserID"
      - $ref: "#/components/parameters/OtherUserID"
    put:
      tags: ["UserInteraction"]
      operationId: banUser
      summary: Ban another user
      description : "Hide the user if it's not already hidden, else do nothing"
      responses:
        '204': { description: "User banned successfully." }
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized"}
        '403': { $ref: "#/components/responses/Forbidden" }
        '404': { $ref: "#/components/responses/ResultNotFound" }
        '409': { description: "The user is already banned, cannot proceed" }
        '500': { $ref: "#/components/responses/InternalServerError" }
    delete:
      tags: ["UserInteraction"]
      operationId: unbanUser
      summary: Unban another user
      description : "Unhide the user if it's hidden, else do nothing"
      responses:
        '204': { description: "User unbanned successfully" }
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized"}
        '403': { $ref: "#/components/responses/Forbidden" }
        '404': { $ref: "#/components/responses/ResultNotFound" }
        '500': { $ref: "#/components/responses/InternalServerError" }
...
  /photos/{PhotoID}:
    parameters:
      - $ref: "#/components/parameters/PhotoID"
    get:
      tags: ["ManagePhotos"]
      operationId: getPhotoInfo
      summary: Get the photo's info
      description : "Get the photo's info"
      responses:
        '200':
          description: "Photo info retrieved successfully." 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo" 
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized"}
        '403': { $ref: "#/components/responses/Forbidden" }
        '404': { $ref: "#/components/responses/ResultNotFound" }
        '500': { $ref: "#/components/responses/InternalServerError" }
    delete:
      tags: ["ManagePhotos"]
      operationId: deletePhoto
      summary: Delete photo
      description : "Delete a photo owned by the user"
      responses:
        '204': { description: "Photo deleted." }
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized"}
        '403': { $ref: "#/components/responses/Forbidden" }
        '404': { $ref: "#/components/responses/ResultNotFound" }
        '500': { $ref: "#/components/responses/InternalServerError" }
  /photos/{PhotoID}/file:
    parameters:
      - $ref: "#/components/parameters/PhotoID"
    get:
      tags: ["ManagePhotos"]
      operationId: getPhotoFile
      summary: Get photo file
      description : "Get the photo file"
      responses:
        '200':
          description: The photo has been successfully retrieved
          content:
            image/*:
              schema:
                type: string
                format: binary
                description: The requested photo
                pattern: '[\s\S]+'
                minLength: 1
                maxLength: 18446744073709551615
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized"}
        '403': { $ref: "#/components/responses/Forbidden" }
        '404': { $ref: "#/components/responses/ResultNotFound" }
        '500': { $ref: "#/components/responses/InternalServerError" }
...
  /photos/{PhotoID}/comments:
    parameters:
      - $ref: "#/components/parameters/PhotoID"
    post:
      tags: ["UserInteraction"]
      operationId: commentPhoto
      summary: Comment a photo
      description : "Comment a photo"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: "Object ontaining the comment"
              properties:
               Comment:
                  $ref: "#/components/schemas/CommentContent"
      responses:
        '201':
          description: "Commented the photo"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment" 
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized"}
        '403': { $ref: "#/components/responses/Forbidden" }
        '404': { $ref: "#/components/responses/ResultNotFound" }
        '500': { $ref: "#/components/responses/InternalServerError" }
...
components:
  schemas:
...
    UserProfileSimplified:
      description: "User profile containing the username with the ID"
      type: object
      properties:
        UserID:
          $ref: '#/components/schemas/UserID'
        Username:
          $ref: '#/components/schemas/Username'
...
    CommentContent:
      type: string
      minLength: 1
      maxLength: 60000
      description: The comment
      pattern: '[\s\S]+'
    Username:
      type: string
      description: "The user's username"
      example: Samsey
      pattern: '^.*?$'
      minLength: 3
      maxLength: 16
    Photo:
      type: object
      description: Object containing the photo data
      properties:
        PhotoID:
          $ref: "#/components/schemas/PhotoID"
        Owner:
          $ref: "#/components/schemas/UserProfileSimplified"
        CreationDate:
            $ref: "#/components/schemas/Date"
        Comment:
          $ref: "#/components/schemas/CommentContent"
        Comments:
          type: array
          description: "Array containing the comments in ascending order by time of posting"
          minItems: 0
          maxItems: 100000
          items:
            $ref: "#/components/schemas/Comment"
    Comment:
      type: object
      description: Object containing the comment data
      properties:
        CommentID:
          $ref: "#/components/schemas/CommentID"
        Comment:
          $ref: "#/components/schemas/CommentContent"
        Owner:
          $ref: "#/components/schemas/UserProfileSimplified"
        CreationDate:
          $ref: "#/components/schemas/Date"
    UserID:
      type: integer
      description: "User's numerical ID"
      example: 1
    PhotoID:
      type: integer
      description: "Photo's numerical ID"
      example: 1
    CommentID:
      type: integer
      description: "Comment's numerical ID"
      example: 1
    Date:
        description: Date 
        type: string
        format: date-time
        minLength: 3
        maxLength: 20
        pattern: '[\s\S]+'
  parameters:
...
    UserID:
      schema:
        description: Active User ID
        type: integer
        example: 1
        readOnly: true
      name: UserID
      in: path
      required: true
      description: Active User ID for parameter
    OtherUserID:
      schema:
        description: Other User ID
        type: integer
        example: 1
        readOnly: true
      name: OtherUserID
      in: path
      required: true
      description: Other User ID for parameter
    PhotoID:
      schema:
        description: Photo ID
        type: integer
        example: 1
        readOnly: true
      name: PhotoID
      in: path
      required: true
      description: Photo ID for parameter
...
  responses:
    Unauthorized:
      description: The request was not authenticated
    Forbidden:
      description: The request was not authorized
    ResultNotFound:
      description: The requested result has not been found.
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc)
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []